refactor(chat): extract parseSystemMessage into a pure util
Move parseSystemMessage (and its parseIdNameToken helper) verbatim out of
MessageDisplay.jsx into src/utils/messageSystemParser.js. The parser is
pure (no React/JSX) and was already imported by three non-component
callers (eventPreview, MessageNotifications, Chat), so it belongs in a
shared util rather than a component file. Importers now point at the util;
MessageDisplay imports it for its own use.

No behavior change — verbatim move plus import rewiring. MessageDisplay
drops from 5300 to 4597 lines.

Stage 1 of the MessageDisplay.jsx decomposition (Tier 2).
