+++ src/pages/Chat.jsx
@@ -16,7 +16,8 @@ import { fetchTeamById } from "../hooks/useTeamQueries";
 import { useConversations, conversationsQueryKey } from "../hooks/useChatQueries";
 import PageContainer from "../components/layout/PageContainer";
 import ConversationList from "../components/chat/ConversationList";
-import MessageDisplay, { parseSystemMessage } from "../components/chat/MessageDisplay";
+import MessageDisplay from "../components/chat/MessageDisplay";
+import { parseSystemMessage } from "../utils/messageSystemParser";
 import MessageInput from "../components/chat/MessageInput";
 import { useAuth } from "../contexts/AuthContext";
 import { messageService } from "../services/messageService";
